move chan_close_confirm
